Update dependancies, fix schema.prisma, finish auth and users services

+++ src/users/users.service.ts
@@ -1,7 +1,12 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
 import {
-  User,
+  BadRequestException,
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
+import {
   UserCreateInput,
+  UserCreateWithoutPastesInput,
   UserOrderByInput,
   UserUpdateInput,
   UserWhereInput,
@@ -14,8 +19,17 @@ import { PrismaService } from '../prisma/prisma.service';
 export class UsersService {
   constructor(private readonly prisma: PrismaService) {}
 
-  async findOne(where: UserWhereUniqueInput): Promise<User | null> {
+  async findOne(where: UserWhereUniqueInput) {
     return this.prisma.user.findOne({ where });
+  }
+
+  async findOneOrCreate(where: UserWhereUniqueInput, data: UserCreateInput) {
+    const user = await this.prisma.user.findOne({ where });
+    if (user) {
+      return user;
+    }
+
+    return this.prisma.user.create({ data });
   }
 
   async findMany(params: {
@@ -28,7 +42,7 @@ export class UsersService {
     return this.prisma.user.findMany(params);
   }
 
-  async createUser(data: UserCreateInput) {
+  async createUser(data: UserCreateWithoutPastesInput) {
     const user = await this.prisma.user.findOne({
       where: { email: data.email },
     });
@@ -36,20 +50,34 @@ export class UsersService {
       throw new BadRequestException('User already exists');
     }
 
-    const password = await bcrypt.hash(data.password, 10);
+    const password = await bcrypt.hash(data.password!, 10);
     return this.prisma.user.create({ data: { ...data, password } });
   }
 
-  async updateUser(params: {
-    where: UserWhereUniqueInput;
-    data: UserUpdateInput;
-  }) {
-    // TODO: check if has rights
-    return this.prisma.user.update(params);
+  async updateUser(
+    where: UserWhereUniqueInput,
+    data: UserUpdateInput,
+    currentUserId: number,
+  ) {
+    const user = await this.prisma.user.findOne({ where });
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+    if (user.id !== currentUserId) {
+      throw new ForbiddenException();
+    }
+    return this.prisma.user.update({ where, data });
   }
 
-  async deleteUser(where: UserWhereUniqueInput) {
-    // TODO: check if has rights
+  async deleteUser(where: UserWhereUniqueInput, currentUserId: number) {
+    const user = await this.prisma.user.findOne({ where });
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+    if (user.id !== currentUserId) {
+      throw new ForbiddenException();
+    }
+
     return this.prisma.user.delete({ where });
   }
 }
